crew_implementation/pyproject.toml
[project]
name = "crew_implementation"
version = "0.1.0"
description = "crew-implementation using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "openpyxl>=3.1.5",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=0.8.11",
]

[project.scripts]
crew_implementation = "crew_implementation.main:run"
run_crew = "crew_implementation.main:run"
train = "crew_implementation.main:train"
replay = "crew_implementation.main:replay"
test = "crew_implementation.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
